prov/psm,psm2: Make src_addr format consistent with other addresses
The information stored in the "src_addr" field of fi_info is different
from a regular endpoint address. While this address still cannot be
used as input to fi_av_insert, it is helpful to make all the addresses
the same size and be distighuishable from each other. It also makes
it easier to stringfy the "src_addr" for debugging purpose.

Signed-off-by: Jianxin Xiong <jianxin.xiong@intel.com>