prov/psm2: Allow fi_getinfo to take "node" in string address format
If the "node" parameter passed to fi_getinfo is in the string
address format "fi_addr_psmx2://x:y" it is taken directly as the
native address. No further address resolution is performed and
the "service" parameter is ignored.

Otherwise, "node" is interpreted as host name and used together
with "service" for address resolution.

Signed-off-by: Jianxin Xiong <jianxin.xiong@intel.com>