prov/util: Move atomic support to utility code
Provide a set of generic macros for implementing atomic
operations.  These are loosely based on the atomic
implementations in the bgq and psmx2 providers.

Rework the complex data type abstraction to include
missing functionality and align the macros closer to the
macros used by the atomics.

Relocate and rename fi_datatype_size to new ofi_atomic
header file.

The socket provider is updated to use the utility atomics.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>