multi_ep: General code cleanups
Convert buffer pointers from int* to char*.  The buffers are byte
arrays, not integer arrays.  This avoids unnecessary casting.

Clarify that test is not 'pingpong'.  Messages are sent in
both directions, but not in a send-reply fashion.

Use shared fabtest version definition

Restructuring of memory allocations to a central function.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>