usnic: fixed handling of new flags from 1.5.
This commit is a fix to a PR #3036 which updated the provider to
version 1.5, but contained bugs where the provider reset given
mode bits and caused undesired behavior.

This commit added more comments to explain what the code does.

This commit also provides backward compatibility with applications that
compile with 1.5 but request an older API version.

Signed-off-by: Thananon Patinyasakdikul <apatinya@cisco.com>