prov/sockets: No need to check that the ring buffer is empty for TX PE entries
When a remote peer unexpectedly disconnects, it's most likely that some data
in the TX communication ring buffer is waiting to be flushed out.
These buffers cannot however be flushed out if the socket is closed, so there
is no need to wait for the TX ring buffer to be empty, otherwise the
connection will never be marked as disconnected.

This patch also adds the re-initialization to 0 of the ring buffer when
the Progress Entry is released.

Signed-off-by: Sylvain Didelot <sdidelot@ddn.com>
Signed-off-by: Sean Hefty <sean.hefty@intel.com>