prov/sockets: prevent a disabled endpoint from being polled by CQ
When an endpoint is created, it's automatically added to the list
of endpoints that need to be checked for completion.
The problem is that an endpoint can be progressed while it's not
ready yet, leading to undefined structures from being used
(especially in the context of shared received contexts, situation
where I hit the bug).

With this patch, the code for CQ only checks an endpoint when it is
enabled. Also the patch delays the enabling of endpoints to be sure
that all the structures are ready to be polled by the CQ.

Signed-off-by: Sylvain Didelot <sdidelot@ddn.com>