prov/rxm: [cherry-picked] Update data progress type to auto
With 2faac2b505 auto progress is possible in most cases. This patch handles
missing progress on fi_inject case.

On large message sends an app doing an sread or waiting on the CQ file descriptor
may not get a completion when reading the CQ after being woken up from the wait.
The app has to do sread or wait on the file descriptor again.

Signed-off-by: Arun C Ilango <arun.ilango@intel.com>