prov/sockets: prevent the same CQ event from being reported multiple times
I just hit the issue where a unexpected connection tear-down occurs and
2 CQ events with the same context are reported twice.

When an RDMA communication finishes, the corresponding pe_entry is not
immediately removed from the RX list because a response is sent to the
remote peer. At this moment, the CQ event corresponding to the RDMA
completion has been added to the CQ.
If a tear-down occurs while the response is outstanding, the
pe_entry will be canceled and an error event will be added to the CQ
(because the pe_entry is still in the RX list).

This patch now checks if the pe_entry has already been reported in the CQ
before adding the CQ event.

Also memset to 0 pe_entry in sock_pe_progress_buffered_rx() to avoid
access garbage data.

Signed-off-by: Sylvain Didelot <sdidelot@ddn.com>