fabric/fence: Increase strength of FI_FENCE definition
Fixes #3186

The FI_FENCE definition guarantees ordering of the fenced operation
against operations posted prior to the fenced operation.  However,
the behavior of operations posted after the fenced operation is
not defined.  It is unclear if such operations can pass the
fenced operation or operations posted prior to the fencing.

Clarify that the fencing applies to the entire queue of
operations.  Anything else renders the fence operation
essentially meaningless.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>