Signed-off-by: Paul Coffman <pcoffman@anl.gov>
prov/bgq: fi_enbable failing for transmit or receive context on scalable endpoint

The fi_enable for a transmit context on a scalable endpoint was failing
because of the implementation specifics whereby the code was verifying the
existence of a recv_cq if the rx->caps happened to be set or unitiailized,
the fix is to clear the rx->caps for a transmit context on a scalable endpoint.
Vice-versa for the receive context.  Also add extra trace information which
helped identify the issue.