Idea for C++ binding of libfabric.
A very straight-forward binding of libfabric "objects" to
C++ classes with the appropriate fi_* methods minus the
explicit passing of fid_* structs and the fi_ prefix.
See the test.cpp for an example, compile like:
g++ -std=c++11 util/test.cpp -Iinclude -lfabric

This should be able to compile down to the equivalent C implementation /
efficiency-- but I have not checked.

Signed-off-by: pmmccorm <patrick.m.mccormick@intel.com>