prov/sockets: new connection manager for MSG endpoints to limit num of threads
With the current implementation of the connection manager (CM) for
MSG endpoints, one POSIX thread is created for each connection
on both sides. The code also opens many file descriptors because
a socket pair is created to signal any CM thread blocked
in epoll_wait(). Finally, this implementation largely impacts the
scalability of libfabric with the sockets provider and MSG endpoints.

In a nutshell, the patch defines a new structure called
'sock_ep_cm_head', which is the head structure for the new CM.
Each CM head creates 1 polling thread that is responsible for
progressing multiple connection establishments (active and passive
sides).

For now, 1 CM head is created per domain and per Passive Endpoint (PEP).
We could in theory move the CM head from the PEP to the domain, so that
we could keep one single CM head in the domain. However and since a PEP
is not bound to a specific domain (there is no reference kept to the
domain in the structure 'sock_pep'), I couldn't find an elegant way to do it.

Note: this patch doesn't impact the wire protocol and is independent
from the previous patch "prov/sockets: create one listener thread per
domain" although they address the same issue.

Change-Id: I45fe0bf3a87995a0bd191ffc556511aeb6f5f536
Signed-off-by: Sylvain Didelot <sdidelot@ddn.com>