prov/psm, sockets, util: Fix corruption reporting CQ errors to 1.4 users
Fixes #3226

The size of the fi_cq_err_entry structure changed between the 1.4
and 1.5 API.  The updated structure includes a new data field.

The psm, psm2, socket, and util code copy the cq error data using
the size of the 1.5 structure.  As a result, an extra 8 bytes of
data may be written beyond the end of the application's structure.
This has produced stack corruptions, as seen in the
fi_rdm_tagged_peek fabtest when built against 1.4, but executed
against libfabric 1.5.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>