prov/psm2: Fix a segfault when the provider is loaded dynamically
Under certain condition, putenv() is called to add an environment
variable to automatically turn on the PSM2 multi EP support. This
works fine fine as a built-in provider. However, if the provider
is dynamically loaded, the environment becomes corrupted after the
provider is unloaded. This can cause the PSM2 library destructor
to segfault within a getenv() call upon exit() being called.

This is fixed by replacing putenv() with setenv(), which makes a
copy of the string arguments.

Signed-off-by: Jianxin Xiong <jianxin.xiong@intel.com>