src/fabric: Fix to enforce prov ordering on register of DL providers
DL providers are not correctly registered in the order
defined in src/fabric.c for known providers.

This results in unexpected providers being loaded that
may not be the preferred or fastest provider.

To resolve this issue, the registered provider list is now
preconfigured with the known providers and their ranking
such that the known providers are always registered in
the proper order.

Signed-off-by: Spruit, Neil R <neil.r.spruit@intel.com>