prov/util,rxm: Allow self-connection for an endpoint
Issue:
When two endpoints inititiate connection requests to each other at the same
time we have to drop one of those connections to keep memory usage low. We
compare the remote endpoint's listening address with that of local endpoint
to decide which connection to keep. The listening address is also the one used
to do a lookup in the AV table / cmap for fetching the corresponding connection
object.

Some middleware have communication patterns where an endpoint tries to talk to
itself. This causes an issue with the above address comparison as the addresses
will be the same.

Fix:
To avoid this issue only compare the actual address that's used to initiate the
connection. We would still use the listening endpoint's address for AV table /
cmap lookups.

Signed-off-by: Arun C Ilango <arun.ilango@intel.com>