prov/util, rxm: Allow connection to self
Issue:
When two endpoints inititiate connection requests to each other at the same
time we have to drop one of those connections to keep memory usage low. We
compare the remote endpoint's listening address with that of local endpoint
to decide which connection to keep. The listening address is also the one used
to do a lookup in the AV table / cmap for fetching the corresponding connection
object.

Some middleware have communication patterns where an endpoint tries to talk to
itself. This causes a collision in the AV table / cmap lookup.

Fix:
To avoid this issue just create the connection object in the peer list instead
of doing a AV / cmap lookup.

Signed-off-by: Arun C Ilango <arun.ilango@intel.com>