prov/psm2: Fix a potential data corruption in completion generation
When generating completion for a large size RMA that is carried out
as special tagged message, the RMA request is put back into the free
pool before the completion is generated. If the request is reused
soon enough, the completion generation may access the "fi_context"
field with modified content.

Fix this issue by delaying the action of putting back the RMA request
into the free pool.

Signed-off-by: Jianxin Xiong <jianxin.xiong@intel.com>