prov/psm2: Fix long RMA protocol tag collisions
The long RMA protocol uses psm2_mq_i{send,recv}2 to transfer the
payload and uses an active message to tell the destination node to either
post or send a buffer of a certain size. The protocol is currently using the
address of the active message request structure as a unique identifier in
the MQ operation tag. However there is no guarantee that this address will
be unique across nodes, which can lead to tag collisions which may cause the
RMA operation to fail with -FI_ETRUNC.

Take for example two nodes A and B.
-Node A calls fi_write() with B as the destination with a 512 MB payload
--Node A allocates an AM structure with address 0xDEADBEEF
--Node A calls psm2_mq_isend2 with a 512MB buffer with 0xDEADBEEF packed in the tag
--Node A sends the AM to B telling it to post a 512 MB buffer
-Node B simulateously calls fi_read() with A as the destination with a 128 MB payload
--Node B allocates an AM structure with address 0xDEADBEEF (by coincidence)
--Node A calls psm2_mq_irecv2 with a 128 MB buffer with 0xDEADBEEF packed in the tag
--Node B then receives the AM from node A with the instructions to post the
  512 MB buffer.

In the above example, a truncation error will be returned for the fi_write()
call since the buffer posted by node B's fi_read() call will have a matching
tag but is too small to fit the data.

This commit fixes this by setting a bit in the tag to indicate whether or
not the originating operation is a read or a write, which will make the tags
unique in an interleaved fi_write/fi_read situation as described above.

Signed-off-by: Erik Paulson <erik.r.paulson@intel.com>