util: subset attr checking on creation of context
Contexts that are created on a scalable endpoint can only
specify a subset of capabilities as supported by the base
endpoint.

The current sockets provider was merely checking if the
context created was supported by the provider, and not if
it was supported by the base endpoint.

The ofi utility code now expands the appropriate capabilities
and checks against the requested caps.

Signed-off-by: Sayantan Sur <sayantan.sur@intel.com>