prov/psm: Eliminate psm2-compat library delay with hfi devices missing
The PSM2 library may introduce a 15 second delay at device initialization
time if the hfi devices are missing. This has been handled in the psm2
provider by checking the existence of the device files before initializing
the device.

The psm provider didn't handle this situation because the issue doesn't
exist with native PSM library over TrueScale. However, when PSM is
supported via the psm2-compat library over PSM2, the same delay can be
observed.

Now add the same mechanism to the psm provider.

Signed-off-by: Jianxin Xiong <jianxin.xiong@intel.com>