prov/sockets: create one listener thread per domain
With the current implementation, libfabric requires the creation
of many POSIX threads. The also opens opens many file descriptors
because a socket pair is created to signal each connection thread
blocked in epoll_wait().
The current implementation largely impacts the scalability of libfabric
with the socket provider.

Function sock_conn_listen() is responsible for listening for new
connection requests for all kinds of endpoints (MSG, RDM, DGRAM).
The problem is that a thread is created every-time the function
sock_conn_listen() gets executed.
This implementation is particularly inefficient when a server needs
to reply many to connections requests from many clients.

The patch fixes the scalability issue figured out in sock_conn_listen()
by adding one listener thread per OFI domain. This thread is
then responsible for listening connection requests for all EP that
are bound to this specific domain.

Signed-off-by: Sylvain Didelot <sdidelot@ddn.com>