prov/gni: Incorrect hints given to domain check fn
This commit amends the GNI provider behavior with respect to the
ofi_check_domain_attr function and the returned info structure.
The provider was checking the contents of the wrong info
structure and had assumed that the contents were formatted correctly
for the ofi_check_domain_attr function. This commit addresses the
issue by shifting the relevant code to a later point in the
function and changes the fi_info structure to verify from 'hints'
to 'gnix_info'.

Signed-off-by: James Swaro <jswaro@cray.com>