prov/psm2: Add option to build with PSM2 source included
Compiling the source code of PSM2 library together with the provider
allows more aggressive compile-time/link-time optimization. In addition,
the provider can be further optimized by having access to the internal
definitions of the PSM2 library.

This option is enabled with the configure option "--with-psm2-src=DIR".
The resulting provider no longer uses or depends on the stand-alone PSM2
library.

Signed-off-by: Jianxin Xiong <jianxin.xiong@intel.com>