prov/psm2: Add sample script for building the provider with PSM2 source
A sample script is provided for building the psm2 provider with three
different options: linking against libpsm2.so, building with PSM2 source
code compiled in, and building with PSM2 source compiled in and with
link-time optimization enabled.

The "-msse4.2" is already part of the CPPFLAGS for the provider build.
However, due to the use of convenience library, when the LTO option is
used the built-in functions are re-processed at link time and thus the
"-msse4.2" flag needs also to be passed via CFLAGS.

Signed-off-by: Jianxin Xiong <jianxin.xiong@intel.com>