prov/sockets: add support for TCP keepalive for MSG EPs
The current patch enables the support for TCP keepalive for the sockets
provider. Keepalive is particularly interesting in case the remote peer
unexpectedly crashes and cannot send the TCP RST flag to drop the
connection.

By default the support of keepalive is disabled but can be switched on
using the 4 following environment variables:

* FI_SOCKETS_KEEPALIVE_ENABLE:
A boolean to enable the keepalive support.

* FI_SOCKETS_KEEPALIVE_TIME:
An integer to specify the idle time in seconds before sending the first
keepalive probe. Only relevant if FI_SOCKETS_KEEPALIVE_ENABLE is enabled.

* FI_SOCKETS_KEEPALIVE_INTVL:
An integer to specify the time in seconds between individual keepalive probes.
Only relevant if FI_SOCKETS_KEEPALIVE_ENABLE is enabled

* FI_SOCKETS_KEEPALIVE_PROBES: An integer to specify the maximum number
of keepalive probes sent before dropping the connection.
Only relevant if FI_SOCKETS_KEEPALIVE_ENABLE is enabled

Signed-off-by: Sylvain Didelot <sdidelot@ddn.com>