prov/gni: use warning level for kdreg open
In working with libfabric/GNI provider consumers that
are also being used within a Cray MPI application,
opening a domain would fail owing to Cray kdreg
limitations.  It was not obvious from the output
of FI_LOG_LEVEL warn what was going on in this case.

Improve the warn level output from the part of the
GNI provider which is opening kdreg for use by
its rcache.

Signed-off-by: Howard Pritchard <howardp@lanl.gov>