prov/psm2: Allow completion suppression when fi_context is non-NULL.
Previously, when the application set FI_SELECTIVE_COMPLETION and did
not enable FI_COMPLETION but the fi_context is not a NULL pointer,
the provider would enable a completion.  With this change, the
completion is suppressed.

It's possible that an application might want to use the fi_context as
a key to cancel an outstanding operation, but the man pages say that
providers ought to ignore fi_context if completions are supressed,
and so fi_cancel shouldn't be supported in that case.

Signed-off-by: Jim Snow <jim.m.snow@intel.com>