prov/psm2: Improve resource counting for Tx/Rx contexts
Keep track of both the number of contexts supported on the system (the "max"
value), and the number of contexts that are available to the application (the
"free" value). Use them at different places where previously only the "free"
value was used.

A few issues have been addressed with this change:

(1) Premature "out of resource" error due to the "free" value being used
as the context allocation limit. The value is updated with each call to
fi_getinfo and can get smaller after some endpoints have been opened.

(2) Insufficient resource allocation could happen for the same reason when
the "free" value is used to determine the allocation size.

(3) Unable to utilize the last context for non-SEP runs due to different
interpretation of the limit for regular endpoints and scalable endpoints.

Signed-off-by: Jianxin Xiong <jianxin.xiong@intel.com>
(cherry-picked from commit cfeb48fee50eb50cd49cb24ab4f4d5790b00bd83)