Fix manpages have bad whatis entries
Each manual page should start with a "NAME" section, which lists the
name and a brief description of the page separated by "\-". The "NAME"
section is parsed by lexgrog and used to generate a database that's
queried by commands like apropos and whatis. This tag indicates that
lexgrog was unable to parse the NAME section of this manual page.

For manual pages that document multiple programs, functions, files, or
other things, the part before "\-" should list each separated by a
comma and a space. Each thing listed must not contain spaces; a man
page for a two-part command like "fs listacl" must use something like
"fs_listacl" in the "NAME" section so that it can be parsed by
lexgrog.

Refer to the lexgrog(1) manual page, the groff_man(7) manual page, and
the groff_mdoc(7) manual page for details.

Signed-off-by: Benjamin Drung <bdrung@debian.org>