util: Add MR monitor
The MR monitor is a base implementation for memory management
notifications.  It will be used by a memory registration cache
in order to receive notifications of memory regions that are
modified.  For example a monitor region may be freed by the
application or the underlying pages modified.  The MR monitor
will report such changes to the user.

As a base implementation, the MR monitor doesn't do anything
other than define the function prototypes for what will become
the 'real' monitors.

Signed-off-by: Dmitry Gladkov <dmitry.gladkov@intel.com>
Signed-off-by: Sean Hefty <sean.hefty@intel.com>