prov/psm2: Remove the virtual lane mechanism
The virtual lane mechanism was used to emulate multiple logical
send/recv contexts over a single PSM2 endpoint. It provided a useful
workaround for the single-endpoint limitation in older versions of
the PSM2 library.

With the introduction of multi-ep support in the PSM2 library, the
virtual lane mechanism becomes obsolete. Removing the related code
produces a cleaner base line for mapping OFI endpoints directly to
PSM2 endpoints.

Signed-off-by: Jianxin Xiong <jianxin.xiong@intel.com>