prov/sockets: plug an edge case for fi_av_remove
Turns out that there's an edge case in the sockets provider
fi_av_remove method where a socket conn in the table may
be in a connecting state.  This condition needs to be
checked for in order to avoid segfaults.

This behavior was observed highly irregularly (usually took
many 100s of runs) of Open MPI's IBM test suite.  Without
this patch, every few hundred runs on 4 processes one would
hit segfauls with this traceback:

```
(gdb) bt
```

With this patch, the problem is no longer observed.

Signed-off-by: Howard Pritchard <howardp@lanl.gov>