common/shared: Added optional address exchange and sync with sockets
-Added support for configuring an out of band sockets connection.
Configuration of the port used and if the out of band connection is
initalized and used has been added to the ft_opts:
opts->oob_sync -> set to 1 to enable oob socket sync.
opts->oob_port -> set to the port that will be used in the out of band
socket connection.

-Added support for setting the new ft_ops oob_sync & oob_port from the
command-line:
-b enable_oob -> enables oob socket sync on the default port.
-b enable_oob=<port> -> enables oob socket sync on a specified port.

-The out of band socket is used for passing the endpoint
addresses between the host and client during address vector init.

-The out of band socket is also used in ft_sync to perform the test sync
without relying on the provider under test.

-Updated the inital_recv during ep init to check if FI_MSG or FI_TAGGED
capability is set before attempting to post the inital recv.

-Added ft_exchange_addresses_oob to exchange the endpoint addresses
between client and server and add to the address vector.

-Updated av_init_dst to increment the rx_seq to account for
RMA/ATOMIC events given FI_MSG and FI_TAGGED capabilities are not set.

-The out of band socket allows for tests that do not require
FI_MSG/FI_TAGGED capability to be set with the common framework.

Signed-off-by: Spruit, Neil R <neil.r.spruit@intel.com>