prov/psm2: Simplify the handling of self-targeted RMA operations
With the removal of virtual lanes, it is no longer possible for self-
targeted operations to have different endpoints as the initiator and
the target. Remove the duplicated parameter.

Also get the domain pointer from the associated Tx/Rx context instead
of using "fabric->active_domain", allowing this part of code to work
with more than one domain in the future.

Signed-off-by: Jianxin Xiong <jianxin.xiong@intel.com>