getinfo_test.c: Remove reset of FI_PROVIDER.
The original code had an issue if FI_PROVIDER was not in the user
provided environment. The reset code would set the FI_PROVIDER value to
the string "(null)" which would cause providers to be skipped when
trying to run a limited subset of the tests (by commenting out the no
hint tests). Instead of fixing the bug, this PR removes the reset
altogether. The reset is not effective because libfabric creates the
filter at library initialization time (first call right after the test
exports FI_PROVIDER). The rest of the tests will be setting the provider
name in the hints anyway.

Signed-off-by: Ben Turrubiates <bturrubi@cisco.com>