prov/hook: Add framework for hooking all provider calls
Introduce a new provider capable of hooking all calls to a
provider.  The intercepting provider can be used to enable debugging or
performance profiling over any provider.  Because the hook provider
is always built, it is available even in release builds.

Only one hook provider can be enabled at a time.  The selection of
the hook provider is controled through the FI_HOOK environment
variable.  When set, the specified hook is enabled, with all calls
passing through the hook provider before being redirected to the
underlying provider.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>