prov/verbs: handle wildcard addresses correctly
A wildcard fi_info (src_addr) is used for passive endpoints to listen
across all interfaces. In such a case set fabric and domain names to
"any" fabric/domain since the passive endpoint isn't assoicated with
a particular device. At the time of accepting a connection request,
the correct fabric and domain would be identified.

Signed-off-by: Arun C Ilango <arun.ilango@intel.com>