util/ns: Remove calls to ofi_osd_ini/fini
The osd ini/fini calls are made by the library itself, and are
not needed.  Providers that need to call osd ini/fini calls
should do so directly from their own ini/fini routines.  This
is needed because the osd initialization (applies to windows)
may be needed prior to starting the nameserver, and may also
need to be available after the nameserver is stopped.

All current users of the nameserver map to osd ini/fini calls
that are empty.  So, it's safe to simply remove them.

A long term solution is to provide an ini/fini call that all
providers use when built as a DL provider.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>