prov/gni: fixes for UDREG use
Turns out that when libfabric is used with an application that
is also using Cray MPICH, the GNI provider needs to use UDREG if
kdreg support is required.

Unfortunately there was a critical flag left out of the
UDREG attributes modes that caused the GNI provider to
use UDREG without using kdreg.  For applications (like
applications using mercury apparently) that are susceptible
to munmap operations under the covers in glibc's heap allocator,
this can lead to data corruption.

Also, it turns out that the GNI provider needs to be less
aggressive with kdreg entries since the Cray MPICH also uses
up some.

This commit addresses these two issues.

Signed-off-by: Howard Pritchard <howardp@lanl.gov>