prov/psm2: Remove unnecessary dependencies from the configure script
When configuring for building with PSM2 source included, the predefined
macro AC_EGREP_HEADER was used to check for the presence of a required
symbol. Instead of running a straight grep command, the macro would
call cpp to process the file and then run grep on the result. This was
an overkill here and created unnecessary dependencies on all the headers
included in the file being processed.

Use grep directly instead of the macro to check the condition.

Signed-off-by: Jianxin Xiong <jianxin.xiong@intel.com>