prov/psm2: Reduce HW context usage for certain TX only endpoints
Normally a HW context is allocated for each endpoint to be used for
both TX and RX operations. However, if an endpoint uses shared TX
context and doesn't need RX capability there is no need to allocate
a dedicated HW context. This make it possible to reduce the number
of HW contexts needed by using shared TX context.

Signed-off-by: Jianxin Xiong <jianxin.xiong@intel.com>