prov/gni: disabled udreg tests in some cases
Disable the udreg mr cache tests if the GNI provider
has been built with KDREG support.  In this case,
the udreg mr cache doesn't work now that it, itself
is requesting to use the KDREG device.

The underlying problem here is that the KDREG device
has a limitation of one open instance per address
space.  Were this restriction to be lifted by
improvements to KDREG, these tests can be re-enabled.

Signed-off-by: Howard Pritchard <howardp@lanl.gov>