config: Allow custom package location for DSO provider
The "--enable-xxx=PATH" option allows a path be specified for package
"xxx". However, this only works for built-in providers. DSO providers
could only be configured with the "--enable-xxx=dl" option which didn't
have a way to provide a custom path.

Add a new format "--enable-xxx=dl:PATH" so that DSO providers have
the same configuration flexibility as built-in providers.

Signed-off-by: Jianxin Xiong <jianxin.xiong@intel.com>