prov/psm2: Add backward compatibility for earlier PSM2 libraries
The use of psm2_am_register_handlers_2 has allowed a large chunk of
redundant code being removed. However it requires the latest PSM2
library that is currently only available in source format from
https://github.com/01org/opa-psm2 and is not compatible with a
wide range of PSM2 libraries currently in use.

A new scheme is developed to bring back the compatibility without
reintroducing the bulk of the old code removed by this new function.

Now the provider should work with PSM2 release 10.2.235 and later
(corresponding to IFS 10.5 and later).

Signed-off-by: Jianxin Xiong <jianxin.xiong@intel.com>