simple/rdm_rma_*.c: Support OFI_MR_BASIC_MAP correctly
The tests set mr_mode to say that they support OFI_MR_BASIC_MAP,
but assume the key is a known value and that the remote address
is zero. Add ft_exchange_keys() call to exchange key and address,
bump number of counts to wait for. Delete confusing and
overwritten setting of mr_mode to FI_MR_SCALABLE. Fix minor typo
in help messages.

Signed-off-by: John Byrne <john.l.byrne@hpe.com>