streaming/rdm_atomic.c: Fix access flags on buffers
If a provider requires FI_MR_LOCAL, then the three registered
buffers need different permissions. The first buffer is used for
send/recv and local/remote RMA, so it is easier to give it all
permissions. If the results buffer is written by hardware, it
needs FI_READ for the local HCA and, possibly, FI_REMOTE_WRITE
for the remote HCA; similarly the compare buffer needs FI_WRITE
for local and FI_REMOTE_READ for remote. I don't know of a fabric
that would actually access the results/compare data remotely ---
or with hardware for that matter --- but I'm leaving the
possibility open.

Signed-off-by: John Byrne <john.l.byrne@hpe.com>