man: Remove confusing/contradictory text binding an EP to counters
The fi_ep_bind documentation indicates that binding an EP to a CQ
is mandatory (in order to report errors).  However, later the man
page states that an EP may 'instead' be bound to a counter.

The intent is that CQs should always be available in order to provide
details on errors.  Fabtests adheres to this requirement.  Strike the
confusing 'instead of' text.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>