prov/psm2: Update makefile and configure script
All the changes are related to building the provider with PSM2 source
included:

(1) Use the internal uuid implementation and remove libuuid from the
    dependencies. This is consistent with how the PSM2 library is built.

(2) In addition to checking the existence of hfi1_user.h, check that
    the header file has the correct version.

(3) Check both the header file and library of the numactl-devel package.

(4) Provide more descriptive output for failed checks.

(5) Add configure option "--with-numa=DIR" to allow non-standard location
    of the numactl-devel package.

Signed-off-by: Jianxin Xiong <jianxin.xiong@intel.com>