prov/shm, util: standardize FI_ADDR_STR usage in util and shm providers
- add check in utility code during address resolution to ADDR_STR to make sure it is in an acceptable FI_ADDR_STR format before copying it over

- in shm provider, if util_getinfo didn't set src_addr or dest_addr (because it wasn't in the correct format, or we had incorrect input), resolve it according to the following established shm provider standards:

(flags & FI_SOURCE) ? src_addr : dest_addr =
	if (node && service)
		= fi_ns://node:service
	if (service)
		= fi_ns://service
	if (node && !service)
		= fi_shm://node
	if (!node && !service)
		= fi_shm://PID
!(flags & FI_SOURCE)
	src_addr = fi_shm://PID

In other words, if the application provides a source and/or destination address in an acceptable FI_ADDR_STR format (contains "://"), the call to util_getinfo will successfully fill in src_addr and dest_addr with the provided input. If the input is not in an ADDR_STR format, the utility code will leave it unresolved and allow the parent provider (in this case the shared memory provider) to fill it in.
The shared memory provider will then create a proper FI_ADDR_STR address with either the "fi_ns://" (node/service format) or "fi_shm://" (shm format) prefixes signaling whether the addr is a "unique" address and does or does not need an extra endpoint name identifier appended in order to make it unique. For the shared memory provider, we are assuming that the service (with or without a node) is enough to make it unique, but a node alone is not sufficient. If only a node is provided, the fi_shm:// is used to signify that it is not a unique address. If no node or service are provided (and in the case of setting the src address without FI_SOURCE and no hints), the process ID will be used as a default address.
On endpoint creation, if the src_addr has the "fi_shm://" prefix, the provider will append ":[dom_idx]:[ep_idx]" as a unique endpoint name (essentially, in place of a service). In the case of the "fi_ns://" prefix (or any other prefix if one was provided by the application), no supplemental information is required to make it unique and it will remain with only the application-defined address. Note that the actual endpoint name will not include the FI_ADDR_STR "*://" prefix since it cannot be included in any shared memory region names. The provider will strip off the prefix before setting the endpoint name. As a result, the addresses "fi_prefix1://my_node:my_service" and "fi_prefix2://my_node:my_service" would result in endpoints of the same name.
The application can also override the endpoint name after creating an endpoint using setname() without any address format restrictions.

Signed-off-by: aingerson <alexia.ingerson@intel.com>